substring

+++ include/get_system_info.h
@@ -9,6 +9,6 @@ typedef struct{
 }cpuInfo;
 
 cpuInfo get_cpu_info();
-int read_cpu_percentage();
+double read_cpu_percentage();
 
 #endif
